+++ src/resources/elements/components/ui-dg-columns.ts
@@ -58,14 +58,14 @@ export class UIDataColumn {
     return this.element.innerHTML + '&nbsp;';
   }
   getValue(value, record) {
-    return this.processValue(value, record) || '&nbsp;';
+    return this.processValue(value, record);
   }
   processValue(value, record) {
     let retVal = '';
     // let value = record[this.dataId];
     if (isFunction(this.value)) value = this.value(({ value, record }));
     if (isFunction(this.display))
-      retVal = this.display(({ value, record })) || '';
+      retVal = this.display(({ value, record }));
     else {
       switch (this.dataType) {
         case 'age': retVal = UIFormat.age(value); break;
@@ -80,7 +80,7 @@ export class UIDataColumn {
         default: retVal = value; break;
       }
     }
-    return retVal;
+    return isEmpty(retVal) ? '&nbsp;' : retVal;
   }
 
   getSummary(summaryRow, summaryValue, data) {
@@ -240,12 +240,19 @@ export class UIDGLink extends UIDataColumn {
   @bindable() glyph;
   @bindable() label;
   @bindable() class = '';
+  @bindable() show = null;
   @bindable() disabled = null;
 
   isDisabled(value, record) {
     if (isFunction(this.disabled)) return this.disabled(({ value, record }));
     if (this.disabled != null) return record[this.disabled];
     return false;
+  }
+
+  isVisible(value, record) {
+    if (isFunction(this.show)) return this.show(({ value, record }));
+    if (this.show != null) return record[this.show];
+    return true;
   }
 
   getGlyph(value, record) {
@@ -293,12 +300,19 @@ export class UIDGButton extends UIDataColumn {
   @bindable() label;
   @bindable() dropdown;
   @bindable() theme: any = 'default';
+  @bindable() show = null;
   @bindable() disabled = null;
 
   isDisabled(value, record) {
     if (isFunction(this.disabled)) return this.disabled(({ value, record }));
     if (this.disabled != null) return record[this.disabled];
     return false;
+  }
+
+  isVisible(value, record) {
+    if (isFunction(this.show)) return this.show(({ value, record }));
+    if (this.show != null) return record[this.show];
+    return true;
   }
 
   getGlyph(value, record) {