+++ src/resources/elements/components/ui-dropdown.ts
@@ -28,11 +28,13 @@ export class UIDropdown {
       if (getParentByClass(evt.target, 'ui-dropdown') == this.element) return true;
       this.element.classList.remove('ui-open');
     });
+    this.obLocale = UIEvent.subscribe('i18n:locale:changed', e => this.localeChanged());
     UIEvent.queueTask(() => this.valueChanged(this.value));
   }
   detached() {
     this.tether.dispose();
     this.obMouseup.dispose();
+    this.obLocale.dispose();
   }
   // unbind() { }
   // end aurelia hooks
@@ -51,6 +53,7 @@ export class UIDropdown {
   private tether;
   private dropdown;
   private obMouseup;
+  private obLocale;
   private selected;
   private glyph = '';
   private display = '';
@@ -74,6 +77,13 @@ export class UIDropdown {
       this.display = this.defaultText;
       this.glyph = '';
     }
+  }
+
+  localeChanged() {
+    UIEvent.queueTask(() => {
+      let it = this.items.find(it => it.value == this.value);
+      if (it) this.display = it.element.innerText;
+    });
   }
 
   disabledChanged(newValue) {