+++ src/resources/elements/inputs/ui-lists.ts
@@ -132,14 +132,14 @@ export class BaseListInput {
     this.original = this.filtered = groups;
   }
 
-  valueChanged(newValue, onBind = false) {
+  valueChanged(newValue, oldValue?) {
     if (!this.isTagInput) {
       let item = _['findChildren'](this.filtered = this.original, 'items', 'value', newValue === null ? '' : newValue);
       this.elValue = item.text;
       if (!this.forceSelect && !this.elValue) this.elValue = newValue === null ? '' : newValue;
       else if (!this.elValue) this.value = '';
 
-      if (onBind && item.model) UIEvent.fireEvent('select', this.element, this.model = item.model);
+      if (!oldValue && item.model) UIEvent.fireEvent('select', this.element, this.model = item.model);
     }
     else {
       let v = (newValue || '').split(',');
@@ -303,12 +303,11 @@ export class BaseListInput {
   <span class="ui-error" if.bind="errors"><ui-glyph glyph="glyph-invalid"></ui-glyph><ul class="ui-error-list"><li repeat.for="err of errors" innerhtml.bind="err"></li></ul></span>
   <input ref="inputEl" value.bind="elValue" autocomplete="off" size="1"
     focus.trigger="fireEvent($event)" blur.trigger="fireEvent($event)" click.trigger="openDropdown($event)"
-    input.trigger="search() & debounce:200" change.trigger="fireEvent($event)"
+    input.trigger="search() & debounce:200" change.trigger="fireEvent($event)" select.trigger="$event.stopPropagation()"
     keydown.trigger="keyDown($event)" placeholder.bind="placeholder" size="1"
     disabled.bind="isDisabled" readonly.bind="!allowSearch || readonly"/>
   <span class="ui-clear" if.bind="clear && value" click.trigger="clearInput()">&times;</span>
   <span class="ui-input-addon ui-dropdown-handle" click.trigger="openDropdown($event, inputEl.focus())"><ui-glyph glyph="glyph-chevron-down"></ui-glyph></span></div>
-  <div class="ui-input-info" if.bind="info" innerhtml.bind="info"></div>
 
   <div class="ui-list-container ui-floating" ref="dropdown">
     <div if.bind="filtered.length==0" class="ui-list-group">\${emptyText}</div>
@@ -318,6 +317,7 @@ export class BaseListInput {
       <span class="\${iconClass} \${item.icon}" if.bind="item.icon"></span>&nbsp;<span innerhtml.bind="item.display"></span></div>
     </template>
   </div>
+  <div class="ui-input-info" if.bind="info" innerhtml.bind="info"></div>
 </template>`)
 @customElement('ui-combo')
 export class UICombo extends BaseListInput {
@@ -379,7 +379,7 @@ export class UICombo extends BaseListInput {
   <span class="ui-error" if.bind="errors"><ui-glyph glyph="glyph-invalid"></ui-glyph><ul class="ui-error-list"><li repeat.for="err of errors" innerhtml.bind="err"></li></ul></span>
   <div class="ui-tag-item" repeat.for="tag of value | split" if.bind="tag!=''"><span innerhtml.bind="getDisplay(tag)"></span><i class="ui-clear" click.trigger="removeValue(tag)">&times;</i></div>
   <input ref="inputEl" value.bind="elValue" autocomplete="off" size="1"
-    focus.trigger="fireEvent($event)" blur.trigger="fireEvent($event)"
+    focus.trigger="fireEvent($event)" blur.trigger="fireEvent($event)" select.trigger="$event.stopPropagation()"
     input.trigger="search() & debounce:200" change.trigger="fireEvent($event)"
     keydown.trigger="keyDown($event)" placeholder.bind="placeholder"
     disabled.bind="isDisabled" readonly.bind="!allowSearch || readonly"/></div>
@@ -485,7 +485,7 @@ export class UITags extends BaseListInput {
   <input ref="inputEl" value.bind="elValue" class="ui-input ui-remove" autocomplete="off"
     focus.trigger="fireEvent($event)" blur.trigger="fireEvent($event)" size="1"
     input.trigger="search() & debounce:200" change.trigger="fireEvent($event)"
-    keydown.trigger="keyDown($event)" placeholder.bind="placeholder"
+    keydown.trigger="keyDown($event)" placeholder.bind="placeholder" select.trigger="$event.stopPropagation()"
     disabled.bind="isDisabled" readonly.bind="true"/>
   <span class="ui-clear" if.bind="clear && value" click.trigger="clearInput()">&times;</span>